moved site/stage to http.frontend
