+++ frontend/http/src/test/java/org/dotwebstack/framework/frontend/http/stage/StageTest.java
@@ -1,12 +1,11 @@
-package org.dotwebstack.framework.stage;
+package org.dotwebstack.framework.frontend.http.stage;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org. junit.Assert.assertThat;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import org.dotwebstack.framework.site.Site;
-import org.dotwebstack.framework.stage.Stage;
+import org.dotwebstack.framework.frontend.http.site.Site;
 import org.dotwebstack.framework.test.DBEERPEDIA;
 import org.junit.Test;
 import org.junit.runner.RunWith;
